Remove HTTPRequestHead extension from AWSRequest
Move the URL extension to AWSClient as that is where it is used
